unlink package file on remove

+++ frontend/views/package.php
@@ -9,7 +9,11 @@ if(!empty($_POST['move_in_group']) && !empty($_POST['move_from_pos']) && !empty(
 }
 if(!empty($_POST['remove_id']) && is_array($_POST['remove_id'])) {
 	foreach($_POST['remove_id'] as $id) {
-		$db->removePackage($id);
+		$package = $db->getPackage($id);
+		$path = PACKAGE_PATH.'/'.$package->filename;
+		if(unlink($path)) {
+			$db->removePackage($id);
+		}
 	}
 	die();
 }
